Split pattern match validation
